Added authorization.

+++ routes/articles.js
@@ -3,27 +3,43 @@ const router = express.Router();
 const Article = require("./../models/article");
 
 router.get("/new", (req, res) => {
-  res.render("articles/new", { article: new Article() });
+  let isAdmin = false;
+  if (!isAdmin) {
+    res.send("It looks like you don't have access to this section.");
+  } else {
+    res.render("articles/new", { article: new Article() });
+  }
 });
 
 router.get("/edit/:id", async (req, res) => {
-  const article = await Article.findById(req.params.id);
-  res.render("articles/edit", { article: article });
+  let isAdmin = false;
+  if (!isAdmin) {
+    res.send("It looks like you don't have access to this section.");
+  } else {
+    const article = await Article.findById(req.params.id);
+    res.render("articles/edit", { article: article });
+  }
 });
 
 router.get("/:slug", async (req, res) => {
+  let isAdmin = false;
   const article = await Article.findOne({ slug: req.params.slug });
   if (article == null) {
     res.redirect("/");
   }
-  res.render("articles/show", { article: article });
+  res.render("articles/show", { article: article, auth: isAdmin });
 });
 
 router.post(
   "/",
   async (req, res, next) => {
-    req.article = new Article();
-    next();
+    let isAdmin = false;
+    if (!isAdmin) {
+      res.send("It looks like you don't have access to this section.");
+    } else {
+      req.article = new Article();
+      next();
+    }
   },
   saveArticleAndRedirect("new")
 );
@@ -31,15 +47,25 @@ router.post(
 router.put(
   "/:id",
   async (req, res, next) => {
-    req.article = await Article.findById(req.params.id);
-    next();
+    let isAdmin = false;
+    if (!isAdmin) {
+      res.send("It looks like you don't have access to this section.");
+    } else {
+      req.article = await Article.findById(req.params.id);
+      next();
+    }
   },
   saveArticleAndRedirect("edit")
 );
 
 router.delete("/:id", async (req, res) => {
-  await Article.findByIdAndDelete(req.params.id);
-  res.redirect("/");
+  let isAdmin = false;
+  if (!isAdmin) {
+    res.send("It looks like you don't have access to this section.");
+  } else {
+    await Article.findByIdAndDelete(req.params.id);
+    res.redirect("/");
+  }
 });
 
 function saveArticleAndRedirect(path) {
